Engine Change
In an effort to reduce the tokens used and extend my free trial, I've changed the engine to the weaker "curie", and the bot will only be using 5 tweets as a prompt rather than 10. This will result in less accurate tweet simulation, but let's be real the bot's just for fun anyway.